# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AutoCango Vehicle Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shipments:
    get:
      operationId: listShipments
      tags: [shipments]
      summary: List shipments (role-scoped)
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ShipmentStatus"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Shipments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentListResponse"
  /shipments/{shipmentId}:
    get:
      operationId: getShipment
      tags: [shipments]
      summary: Get shipment detail
      security:
        - bearerAuth: []
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /shipments/{shipmentId}/status:
    patch:
      operationId: updateShipmentStatus
      tags: [shipments]
      summary: Update shipment status (forwarder only)
      security:
        - bearerAuth: []
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShipmentStatusInput"
      responses:
        "200":
          description: Shipment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /shipments/{shipmentId}/tracking:
    get:
      operationId: listShipmentTracking
      tags: [shipments]
      summary: Get shipment tracking event log
      security:
        - bearerAuth: []
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tracking events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentTrackingListResponse"
  /shipments/{shipmentId}/documents:
    get:
      operationId: listShipmentDocuments
      tags: [shipments]
      summary: List documents for a shipment (role-filtered)
      security:
        - bearerAuth: []
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShipmentDocument"

    post:
      operationId: uploadShipmentDocument
      tags: [shipments]
      summary: Upload a shipping document (forwarder only)
      security:
        - bearerAuth: []
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadShipmentDocumentInput"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDocument"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /shipments/{shipmentId}/documents/{documentId}:
    delete:
      operationId: deleteShipmentDocument
      tags: [shipments]
      summary: Delete a shipping document (forwarder/admin)
      security:
        - bearerAuth: []
      parameters:
        - name: shipmentId
          in: path
          required: true
          schema:
            type: string
        - name: documentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /orders/{orderId}/shipment:
    get:
      operationId: getOrderShipment
      tags: [shipments]
      summary: Get shipment for an order
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shipment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentDetail"
        "404":
          description: No shipment yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Admin / Super Admin ───────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ShipmentStatus:
      type: string
      enum: [awaiting_quote, quote_submitted, quote_accepted, container_booked, vehicle_collected, at_origin_port, loaded_on_vessel, in_transit, arrived_destination_port, customs_clearance, delivered]
    ShippingDocumentType:
      type: string
      enum: [bill_of_lading, commercial_invoice, packing_list, insurance_certificate, export_declaration, arrival_notice]
    ShipmentSummary:
      type: object
      required: [id, orderId, forwarderId, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        freightRequestId:
          type: string
          nullable: true
        forwarderId:
          type: string
        forwarderName:
          type: string
          nullable: true
        vehicleTitle:
          type: string
          nullable: true
        buyerName:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        notes:
          type: string
          nullable: true
        documentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ShipmentTrackingEvent:
      type: object
      required: [id, shipmentId, status, actorId, createdAt]
      properties:
        id:
          type: string
        shipmentId:
          type: string
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        actorId:
          type: string
        actorName:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    ShipmentDetail:
      type: object
      required: [id, orderId, forwarderId, status, tracking, createdAt, updatedAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        freightRequestId:
          type: string
          nullable: true
        forwarderId:
          type: string
        forwarderName:
          type: string
          nullable: true
        vehicleTitle:
          type: string
          nullable: true
        buyerName:
          type: string
          nullable: true
        sellerName:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        notes:
          type: string
          nullable: true
        tracking:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentTrackingEvent"
        documentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ShipmentListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ShipmentTrackingListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentTrackingEvent"
    ShipmentDocument:
      type: object
      required: [id, shipmentId, documentType, objectPath, fileName, uploadedById, visibleToBuyer, visibleToSeller, createdAt]
      properties:
        id:
          type: string
        shipmentId:
          type: string
        documentType:
          $ref: "#/components/schemas/ShippingDocumentType"
        objectPath:
          type: string
        fileName:
          type: string
        uploadedById:
          type: string
        uploaderName:
          type: string
          nullable: true
        visibleToBuyer:
          type: boolean
        visibleToSeller:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UpdateShipmentStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/ShipmentStatus"
        note:
          type: string
          nullable: true
    UploadShipmentDocumentInput:
      type: object
      required: [documentType, objectPath, fileName]
      properties:
        documentType:
          $ref: "#/components/schemas/ShippingDocumentType"
        objectPath:
          type: string
        fileName:
          type: string
        visibleToBuyer:
          type: boolean
          default: true
        visibleToSeller:
          type: boolean
          default: false

    # ── Freight Conversation Schemas ─────────────────────────────────────────────
...